Remove support for examples from judoc
The judoc examples feature is currently unused.

Keeping support for this feature adds a cost to HTML generation. We are
removing this to improve the performance of `juvix html`.

To just render the HTML documentation we only require the scoper result from
the pipeline. To support the examples we need the type checking result.
The cost is significant in larger projects as the pipeline is run for
each import.