Add lcomposition to support (>>) in the stdlib
The stdlib composition function ∘ has fixity `composition` which means
it is right associative.

We will rename ∘ to << in the stdlib and add a new function:

>> {a b c} : (a -> b) -> (b -> c) -> a -> c

for consistency with << this should be left associative.

This is not a breaking change to package-base so we don't need to
increment th version number.